fix: rename componennts and setup separate utils and consts folders
