+++ src/components/film-card-block/template.js
@@ -1,5 +1,5 @@
-import {createTemplateFilmCard} from './film-card.js';
-import {createTemplateBtnShowMore} from './show-more.js';
+import {createTemplateFilmCard} from '../film-card/template.js';
+import {createTemplateBtnShowMore} from '../show-more/template.js';
 
 export const createTemplateFilmCardBlock = (numberOfCards, contentData) => {
   const {title: {text, isHidden}, isShowMore, isExtraBlock} = contentData;
@@ -9,13 +9,16 @@ export const createTemplateFilmCardBlock = (numberOfCards, contentData) => {
   const templateShowMore = isShowMore ? createTemplateBtnShowMore() : ``;
   const classExtraBlock = isExtraBlock ? `films-list--extra` : `films-list`;
 
+  const templateFilmCards =
+    new Array(numberOfCards).fill(createTemplateFilmCard()).join(``);
+
   return `
     <section class="${classExtraBlock}">
       <h2 class="films-list__title ${classHiddenTitle}">
         ${textTitle}
       </h2>
       <div class="films-list__container">
-        ${new Array(numberOfCards).fill(createTemplateFilmCard()).join(``)}
+        ${templateFilmCards}
       </div>
       ${templateShowMore}
     </section>
