+++ src/components/statistics/template.js
