+++ src/consts/index.js
@@ -1,0 +1,10 @@
+export const PosRender = {
+  BEFORE_END: `beforeend`,
+  BEFORE_BEGIN: `beforebegin`
+};
+
+export const CountFilmCards = {
+  COMMON: 5,
+  TOP_RATED: 2,
+  MOST_COMMENTED: 2
+};
