+++ src/utils/index.js
@@ -1,0 +1,5 @@
+import {PosRender} from '../consts/index.js';
+
+export const render = (container, template, place = PosRender.BEFORE_END) => {
+  container.insertAdjacentHTML(place, template);
+};
